requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
python-dotenv
pymysql
fastapi-pagination